Change the top navigation elements
Change the top navigation elements to add RSS, Twitter and GitHub icons.
Replace the Search option with an icon.

To make sure the page is still accessible the anchor tags include
`aria-label` elements with descriptive text that a screen reader will be
able to read, and the image tag has `aria-hidden="true"` to hide the
icon from screen readers. The image tags also have `title` elements to
generate tooltips.