Only include analytics.html when jekyll.environment is production
The avoid spurious data going to google analytics from local testing I
only want to include the analytics.html code snippet when served from
GitHub Pages. By default jekyll.environment is set to 'development' but
when deployed to GitHub Pages jekyll.environment is set to 'production'.
So use this to decide whether or not to include the analytics snippet.